Switch PerformanceTestFixture to use Mimick
After showing that this approach works, this commit actually swaps the
PerformanceTestFixture from using osrf_testing_tools_cpp to using
Mimick.
