+++ include/performance_test_fixture/performance_test_fixture.hpp
@@ -16,12 +16,10 @@
 #define PERFORMANCE_TEST_FIXTURE__PERFORMANCE_TEST_FIXTURE_HPP_
 
 #include <benchmark/benchmark.h>
-#include <osrf_testing_tools_cpp/memory_tools/memory_tools.hpp>
 
-#include <atomic>
+#include <memory>
 
-#include "performance_test_fixture/common.hpp"
-#include "performance_test_fixture/visibility_control.hpp"
+#include "performance_test_fixture/mimick_memory_manager.hpp"
 
 namespace performance_test_fixture
 {
@@ -29,31 +27,39 @@ namespace performance_test_fixture
 class PerformanceTest : public benchmark::Fixture
 {
 public:
-  PERFORMANCE_TEST_FIXTURE_PUBLIC
   PerformanceTest();
 
-  PERFORMANCE_TEST_FIXTURE_PUBLIC
   void SetUp(benchmark::State & state) override;
 
-  PERFORMANCE_TEST_FIXTURE_PUBLIC
   void TearDown(benchmark::State & state) override;
 
 protected:
-  PERFORMANCE_TEST_FIXTURE_PUBLIC
-  void on_alloc(osrf_testing_tools_cpp::memory_tools::MemoryToolsService & service);
-
-  PERFORMANCE_TEST_FIXTURE_PUBLIC
   void reset_heap_counters();
 
-  PERFORMANCE_TEST_FIXTURE_PUBLIC
   void set_are_allocation_measurements_active(bool value);
 
-private:
-  std::atomic_size_t allocation_count;
-  bool suppress_memory_tools_logging;
-  bool are_allocation_measurements_active;
+  std::unique_ptr<MimickMemoryManager> memory_manager;
 };
 
 }  // namespace performance_test_fixture
+
+/**
+ * Macro to pause timing and heap allocation measurements over a section of code.
+ *
+ * This is useful if there is setup or teardown that has to occur with every iteration.
+ * For example, if you wanted to measure only construction of a rclcpp::Node and not its
+ * destruction.
+ *
+ * state.PauseTiming() does not allocate, so it should go first to minimize discrepencies in
+ * measuring timing
+ */
+#define PERFORMANCE_TEST_FIXTURE_PAUSE_MEASUREMENTS(state, code) \
+  do { \
+    state.PauseTiming(); \
+    set_are_allocation_measurements_active(false); \
+    code; \
+    set_are_allocation_measurements_active(true); \
+    state.ResumeTiming(); \
+  } while (0)
 
 #endif  // PERFORMANCE_TEST_FIXTURE__PERFORMANCE_TEST_FIXTURE_HPP_
